+++ open-vm-tools/lib/include/vm_poll.h
